Moved Cassandra test utility methods to a new file so they can be used outside of temporal repo.
