+++ common/persistence/tests/cassandra_test.go
@@ -25,17 +25,13 @@
 package tests
 
 import (
-	"fmt"
-	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/suite"
 
 	"go.temporal.io/server/common/config"
 	"go.temporal.io/server/common/log"
-	p "go.temporal.io/server/common/persistence"
 	"go.temporal.io/server/common/persistence/cassandra"
-	"go.temporal.io/server/common/persistence/nosql/nosqlplugin/cassandra/gocql"
 	_ "go.temporal.io/server/common/persistence/sql/sqlplugin/mysql"
 	"go.temporal.io/server/common/resolver"
 	"go.temporal.io/server/common/shuffle"
@@ -50,17 +46,12 @@ const (
 	testCassandraPassword           = "temporal"
 	testCassandraDatabaseNamePrefix = "test_"
 	testCassandraDatabaseNameSuffix = "temporal_persistence"
-
-	// TODO hard code this dir for now
-	//  need to merge persistence test config / initialization in one place
-	testCassandraExecutionSchema  = "../../../schema/cassandra/temporal/schema.cql"
-	testCassandraVisibilitySchema = "../../../schema/cassandra/visibility/schema.cql"
 )
 
 func TestCassandraExecutionMutableStateStoreSuite(t *testing.T) {
 	cfg := newCassandraConfig()
 	SetupCassandraDatabase(cfg)
-	SetupCassandraSchema(cfg)
+	SetUpCassandraSchema(cfg)
 	logger := log.NewNoopLogger()
 	factory := cassandra.NewFactory(
 		*cfg,
@@ -88,7 +79,7 @@ func TestCassandraExecutionMutableStateStoreSuite(t *testing.T) {
 func TestCassandraHistoryStoreSuite(t *testing.T) {
 	cfg := newCassandraConfig()
 	SetupCassandraDatabase(cfg)
-	SetupCassandraSchema(cfg)
+	SetUpCassandraSchema(cfg)
 	logger := log.NewNoopLogger()
 	factory := cassandra.NewFactory(
 		*cfg,
@@ -112,7 +103,7 @@ func TestCassandraHistoryStoreSuite(t *testing.T) {
 func TestCassandraTaskQueueSuite(t *testing.T) {
 	cfg := newCassandraConfig()
 	SetupCassandraDatabase(cfg)
-	SetupCassandraSchema(cfg)
+	SetUpCassandraSchema(cfg)
 	logger := log.NewNoopLogger()
 	factory := cassandra.NewFactory(
 		*cfg,
@@ -136,7 +127,7 @@ func TestCassandraTaskQueueSuite(t *testing.T) {
 func TestCassandraTaskQueueTaskSuite(t *testing.T) {
 	cfg := newCassandraConfig()
 	SetupCassandraDatabase(cfg)
-	SetupCassandraSchema(cfg)
+	SetUpCassandraSchema(cfg)
 	logger := log.NewNoopLogger()
 	factory := cassandra.NewFactory(
 		*cfg,
@@ -165,87 +156,5 @@ func newCassandraConfig() *config.Cassandra {
 		Hosts:    environment.GetCassandraAddress(),
 		Port:     environment.GetCassandraPort(),
 		Keyspace: testCassandraDatabaseNamePrefix + shuffle.String(testCassandraDatabaseNameSuffix),
-	}
-}
-
-func SetupCassandraDatabase(cfg *config.Cassandra) {
-	adminCfg := *cfg
-	// NOTE need to connect with empty name to create new database
-	adminCfg.Keyspace = "system"
-
-	session, err := gocql.NewSession(adminCfg, resolver.NewNoopResolver(), log.NewNoopLogger())
-	if err != nil {
-		panic(fmt.Sprintf("unable to create Cassandra session: %v", err))
-	}
-	defer session.Close()
-
-	if err := cassandra.CreateCassandraKeyspace(
-		session,
-		cfg.Keyspace,
-		1,
-		true,
-		log.NewNoopLogger(),
-	); err != nil {
-		panic(fmt.Sprintf("unable to create Cassandra keyspace: %v", err))
-	}
-}
-
-func SetupCassandraSchema(cfg *config.Cassandra) {
-	session, err := gocql.NewSession(*cfg, resolver.NewNoopResolver(), log.NewNoopLogger())
-	if err != nil {
-		panic(fmt.Sprintf("unable to create Cassandra session: %v", err))
-	}
-	defer session.Close()
-
-	schemaPath, err := filepath.Abs(testCassandraExecutionSchema)
-	if err != nil {
-		panic(err)
-	}
-
-	statements, err := p.LoadAndSplitQuery([]string{schemaPath})
-	if err != nil {
-		panic(err)
-	}
-
-	for _, stmt := range statements {
-		if err = session.Query(stmt).Exec(); err != nil {
-			panic(err)
-		}
-	}
-
-	schemaPath, err = filepath.Abs(testCassandraVisibilitySchema)
-	if err != nil {
-		panic(err)
-	}
-
-	statements, err = p.LoadAndSplitQuery([]string{schemaPath})
-	if err != nil {
-		panic(err)
-	}
-
-	for _, stmt := range statements {
-		if err = session.Query(stmt).Exec(); err != nil {
-			panic(err)
-		}
-	}
-}
-
-func TearDownCassandraKeyspace(cfg *config.Cassandra) {
-	adminCfg := *cfg
-	// NOTE need to connect with empty name to create new database
-	adminCfg.Keyspace = "system"
-
-	session, err := gocql.NewSession(adminCfg, resolver.NewNoopResolver(), log.NewNoopLogger())
-	if err != nil {
-		panic(fmt.Sprintf("unable to create Cassandra session: %v", err))
-	}
-	defer session.Close()
-
-	if err := cassandra.DropCassandraKeyspace(
-		session,
-		cfg.Keyspace,
-		log.NewNoopLogger(),
-	); err != nil {
-		panic(fmt.Sprintf("unable to drop Cassandra keyspace: %v", err))
 	}
 }
